+++ pyfeng/__init__.py
@@ -19,16 +19,12 @@ from .multiasset import (
     BsmSpreadKirk,
     BsmSpreadBjerksund2014,
     NormBasket,
-    NormSpread,
     BsmBasketLevy1992,
     BsmMax2,
     BsmBasketMilevsky1998,
-    BsmBasket1Bm,
-    BsmBasketLowerBound,
+    BsmBasket1Bm, BsmBasketChoi2018,
     BsmBasketJsu,
 )
-
-from .multiasset_sumbsm import BsmBasketChoi2018
 
 from .multiasset_mc import BsmNdMc, NormNdMc
