+++ pyfeng/opt_abc.py
@@ -487,61 +487,3 @@ class OptAnalyticABC(OptABC):
         return NotImplementedError
 
 
-class OptMaABC(OptABC, abc.ABC):
-
-    n_asset = 1
-    rho = None
-    cor_m = np.diag([1.0])
-    cov_m = np.diag([1.0])
-    chol_m = np.diag([1.0])
-
-    def __init__(self, sigma, cor=None, intr=0.0, divr=0.0, is_fwd=False):
-        """
-
-        Args:
-            sigma: model volatilities of `n_asset` assets. (n_asset, ) array
-            cor: correlation. If matrix with shape (n_asset, n_asset), used as it is.
-                If scalar, correlation matrix is constructed with all same off-diagonal values.
-            intr: interest rate (domestic interest rate)
-            divr: vector of dividend/convenience yield (foreign interest rate) 0-D or (n_asset, ) array
-            is_fwd: if True, treat `spot` as forward price. False by default.
-        """
-        sigma = np.atleast_1d(sigma)
-        self.n_asset = len(sigma)
-
-        super().__init__(sigma, intr, divr, is_fwd)
-
-        if self.n_asset == 1:
-            if cor is not None:
-                print(f"Ignoring cor={cor} for a single asset")
-            self.rho = None
-            self.cor_m = np.array([[1.0]])
-        elif np.isscalar(cor):
-            self.cor_m = cor * np.ones((self.n_asset, self.n_asset)) + (
-                1 - cor
-            ) * np.eye(self.n_asset)
-            self.rho = cor
-        else:
-            assert cor.shape == (self.n_asset, self.n_asset)
-            self.cor_m = cor
-            if self.n_asset == 2:
-                self.rho = cor[0, 1]
-
-        self.cov_m = sigma * self.cor_m * sigma[:, None]
-        self.chol_m = np.linalg.cholesky(self.cov_m)
-
-    def price(self, strike, spot, texp, cp=1):
-        """
-        Call/put option price.
-
-        Args:
-            strike: strike price.
-            spot: spot (or forward) prices for assets.
-                Asset dimension should be the last, e.g. (n_asset, ) or (N, n_asset)
-            texp: time to expiry.
-            cp: 1/-1 for call/put option.
-
-        Returns:
-            option price
-        """
-        return NotImplementedError
